Run each worker over one WorkerLink connection
Control, agent sessions and file transfers become channels on /w, opened by
the server with the request as the header. A dropped connection only
detaches the link: agents keep running and resume on reconnect; a dead link
tears the worker down. The worker's global session, eval-host and update
state moves into a Worker object.

Also: test servers no longer scan the host's agent sessions on connect,
TestKit scrubs every mock knob before a new server, and spawn failures no
longer leak the agent environment into error messages.

+++ BonitoAgents/test/e2e/scroll_anchor_test.jl
@@ -232,7 +232,7 @@
         @test Int(r["after"]) > div(Int(r["before"]), 2)   # pre-fix: 0 (first message)
         # ...and the reading row comes back: the queued refresh un-parks it and
         # the next pin is exact again, at the offset it was captured at.
-        idx = Int(r["idx"]); off = Int(r["off"])
+        idx = Int(r["idx"]); off = Float64(r["off"])   # px, fractional and often negative
         @test TK.wait_for(s, "parked reading row re-materialised at its offset",
             """(() => {
                 const ch = $CH;
